requirements.txt
ccxt==4.5.15
python-telegram-bot==22.5
